-Updated Node server structure. 	-Refine the organization and removed fat. 	-Organized userController in Node.
